Fix failing test checking app initialization
It turns out that the test was missing a lot of context logic required by the <App /> component and was failing because, for example, the store was not passed down to nested components. The redux documentation on testing (http://redux.js.org/docs/recipes/WritingTests.html#connected-components) finally provided the necessary hint.

This test is somewhat useful, in that it checks whether the app can be loaded in a jsdom environment at all. However, because it doesn't test functionality and isn't very specific, its use is limited (problems it uncovers will probably pop up during the build stage). It might therefore be removed at some later point.