Stop pretending we can serialize buttons in a form
The library previously serialized all buttons present in a form, saving name and value attributes as key/value pairs. Standard HTML forms do not do this, but save name and value only for the button that was used to submit the form. This change stops serialization for any button data, to avoid the incorrect assumption that the data stem from the submitted button (it doesn't, buttons are serialized sequentially). At some later point, the code might be extended to capture the data for the submitted button, but that would not be an easy change, and popular serialization libraries don't implement this feature for the stated reason.