Improve introduction to functions, and replace problematic example
This commit removes an example with a sexist connotation from the documentation. The example was added without ill intent and without reflection, not understanding what it conveys. Having been alerted to its history, it is, however, obviously inappropriate and demeaning. The author would like to thank an attentive and honest reader for pointing it out, and apologize for perpetuating an idea that has no place in this project, nor in the world at large for that matter.