Balance the number of congruent and incongruent trials in the Stroop task example
Many thanks to the members of the @PerceptionAndCognitionLab, [Jeff Rouder](https://github.com/rouderj), [Julia Haaf](https://github.com/jstbcs) and [Hope K. Snyder](https://osf.io/7y6bv/) for pointing out this pitfall in their fantastic paper on [Minimizing Mistakes In Psychological Science](https://psyarxiv.com/gxcy5). They point to [Logan & Zbrodoff (1979)](https://dx.doi.org/10.3758/BF03197535), who report an attenuation (and even a reversal) of the Stroop effect for large proportions of incongruent trials. Because effects change across the entire range of task composition, it appears to the writer that there is no easy solution to this: As [MacLeod (1991)](https://dx.doi.org/10.1037/0033-2909.109.2.163) notes, a convincing demonstration of an interference effect should probably include both homogenous and mixed blocks. Thus, though hopefully an improvement, this is probably not the last change the design of this example. Inadequacies, of course, remain the responsibility of the author.