Clone data before saving to remove references to nested objects
Previously, changing objects and arrays in a datastore's state might change previous data which was linked by reference. With this change, data from staging is recursively cloned before being added to the log, breaking any references and ensuring immutability of stored information even if it is encapsuled in a container.

Massive kudos to Michael Kriechbaumer for bringing this behaviour to my attention!