Skip updates to missing components
This fixes an issue that arose in deea1f, where we started flushing pending updates to the canvas editor when unmounting it. Deleting a component removes it from the application state and also unmounts the canvas editor, causing an update to a non-existent component and resulting in an invalid application state. We discussed several different places in which to check for the existence of the data, and this choice is debatable, but it's the simplest; the commit also adds a note to indicate that this might require revisiting.