Include parameters in option parsing / placeholder resolution step
This will allow for parameters with embedded expressions/placeholders that use parameters from superordinate components. Using other options from the same component won't work (for now, we discussed a two-step parsing method that makes local parameters available inside of the other options as a next step).

Many thanks to @nishkakar for suggesting this feature and helping figure out the implementation!