Use relative paths in JATOS export
We previously assumed that a JATOS instance always runs at a top-level URL, and this supports configurations that run in a subdirectory. With this change, we require a minimum JATOS version of 3.3.1.

Many heartfelt thanks to Lea Hildebrandt for identifying this issue, and to Kristian Lange (@kristian-lange, of JATOS fame) for helping with the technical details!