Remove remnants of CI automation via Travis-CI
I'm really sorry to see travis-ci.org go, but we're not going to move to their commercial service. Many thanks to the kind folks at Travis for introducing me to CI, for supporting this project (we were at almost 2000 builds) and for the rainbow mode, which has brought me joy so many times over the past few years.