Revise reset logic
Attempts to more cleanly separate reset logic between components and controller. This is likely not the last iteration for this, as both method naming and placement of logic are in flux.

Specifically:
- Components now expose. a `reset` method, which re-runs a currently active component. This replaces the `rerun` method that previously offered this functionality, but only for the lowest-level components (screens), while this functionality is now more widely supported (there are still some outstanding bugs in canvas components). We might rename the method back to `rerun`.
- Some of the functionality previously contained in `reset` (which was not intended for public use, and only called through the controller) is still contained in a private (to TypeScript) helper method `_reset`. This method is intended to take an already-run component and take it back to a running state. Is a pragmatic choice for the time being, and I'm looking for a better solution -- the code might to one of its consumers, either the controller internals or possibly the `run` method.

It feels like there's an overall API design issue here, in that `prepare`, `run` and `end` are on one hand methods that are called by the study controller to manage the progression of components throughout a study, but also, on the other, called by users who would like to control of this progression (e.g. studies are typically triggered by calling `run` on the top-level sequence/loop, and `end` and `reset` are available on running components).
In the future, it might become necessary to give up this (latter) functionality and decide that components only provide callbacks to be run through the controller, and users looking to control the study flow interact only with the controller.